+++ staily/src/main/java/com/itwill/staily/main/mapper/MainMapper.java
@@ -1,13 +1,10 @@
 package com.itwill.staily.main.mapper;
 
 import java.util.List;
-import java.util.Map;
 
 import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
 
 import com.itwill.staily.mypage.model.dto.Bookmark;
-import com.itwill.staily.util.Member;
 import com.itwill.staily.util.Product;
 import com.itwill.staily.util.Work;
