+++ staily/src/main/resources/com/itwill/staily/main/mapper/ListMapper.xml
@@ -5,9 +5,8 @@
 		<collection property="product" ofType="Product" autoMapping="true">
 		</collection>
 	</resultMap>
-	<!-- 4.작품상세페이지 작품정보(작품이름,작품 포스터) 출력하기 -->
-	<!-- 5-1.작품상세페이지 상품정보(상품이름, 상품에피소드,상품씬,작성자) 출력하기 -->
-	<!-- 5-2.작품상세페이지 상품정보 회차별 출력하기 -->
+	<!-- 2.작품상세페이지 상품정보(상품이름, 상품에피소드,상품씬,작성자) 출력하기 -->
+	<!-- 3.작품상세페이지 상품정보 회차별 출력하기 -->
 	
 	<!-- 4.선택한 작품의 포스터,상품이름 출력하기 -->
 	<select id="selectByWork" parameterType="int" resultMap="wWithpAndpdResultMap">
@@ -33,37 +32,18 @@
 		on p.p_no = pd.p_no
 		where w.w_no = #{wNo} and wd.wd_episode=#{wdEpisode} and p.p_check='Y'
 	</select>
-	<!-- 6.즐겨찾기 등록하기 -->
+	<!-- 3.즐겨찾기 등록 -->
 	<insert id="createBookmark" parameterType="map"> 
 		<selectKey order="BEFORE" keyProperty="bmNo" resultType="int">
 			select BM_NO_SEQ.nextval from dual
 		</selectKey>
 		insert into bookmark values(#{bmNo},#{mNo},#{pNo})
 	</insert>
-	<!-- 7.즐겨찾기 제거하기 -->
+	<!-- 4.즐겨찾기 제거 -->
 	<delete id="deleteBookmark" parameterType="int" >
 		delete from bookmark where bm_no=#{bmNo}
 	</delete>
-	<!-- 8.총회차 찾기 -->
-	<select id="selectTepisode" parameterType="int" resultType="Work">
-		select w_tepisode
-		from work
-		where w_no=#{wNo}
-	</select>
-	<!-- 9.작품 조회수 증가 -->
-	<update id="increaseWorkView" parameterType="int">
-		update work
-		set w_view = w_view + 1
-		where w_no = #{wNo}
-	</update>
-	<!-- 10.상품 조회수 증가 -->
-	<update id="increaseProductView" parameterType="int">
-		update product
-		set p_view = p_view + 1
-		where p_no = #{pNo}
-		and p_check='Y'
-	</update>
-	<!-- 11.게시물 총 건수 -->		
+	<!-- 5.게시물 총 건수 -->		
 	<select id="selectProductCount" resultType="int">
 		select count(*) from product where p_check='Y'
 	</select>
